requirements.txt
django~=3.2
djangorestframework~=3.12
redis~=3.5
django-celery-beat~=2.2
djoser~=2.1